requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
Django==3.0.5
ffmpeg==1.4
gast==0.2.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
joblib==0.14.1
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.1.3
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.2.1
pandas==0.23.3
Pillow==7.1.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
urllib3==1.25.9
webcolors==1.3
Werkzeug==1.0.1
wrapt==1.12.1
